Motion notifications: dedicated, configurable sound (own picker + player)
